Struct aws_sdk_cloudformation::client::fluent_builders::CreateChangeSet

pub struct CreateChangeSet<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateChangeSet<C>

pub async fn send(
self
) -> Result<CreateChangeSetOutput, SdkError<CreateChangeSetError>> where
C: SmithyConnector,

pub fn stack_name(self, input: impl Into<String>) -> Self

The name or the unique ID of the stack for which you are creating a change set. AWS CloudFormation generates the change set by comparing this stack's information with the information that you submit, such as a modified template or different parameter input values.

pub fn set_stack_name(self, input: Option<String>) -> Self

pub fn template_body(self, input: impl Into<String>) -> Self

A structure that contains the body of the revised template, with a minimum length of 1 byte and a maximum length of 51,200 bytes. AWS CloudFormation generates the change set by comparing this template with the template of the stack that you specified.

Conditional: You must specify only TemplateBody or TemplateURL.

pub fn set_template_body(self, input: Option<String>) -> Self

pub fn template_url(self, input: impl Into<String>) -> Self

The location of the file that contains the revised template. The URL must point to a template (max size: 460,800 bytes) that is located in an S3 bucket or a Systems Manager document. AWS CloudFormation generates the change set by comparing this template with the stack that you specified.

Conditional: You must specify only TemplateBody or TemplateURL.

pub fn set_template_url(self, input: Option<String>) -> Self

pub fn use_previous_template(self, input: bool) -> Self

Whether to reuse the template that is associated with the stack to create the change set.

pub fn set_use_previous_template(self, input: Option<bool>) -> Self

pub fn parameters(self, inp: impl Into<Parameter>) -> Self

A list of Parameter structures that specify input parameters for the change set. For more information, see the Parameter data type.

pub fn set_parameters(self, input: Option<Vec<Parameter>>) -> Self

pub fn capabilities(self, inp: impl Into<Capability>) -> Self

In some cases, you must explicitly acknowledge that your stack template contains certain capabilities in order for AWS CloudFormation to create the stack.

CAPABILITY_IAM and CAPABILITY_NAMED_IAM

Some stack templates might include resources that can affect permissions in your AWS account; for example, by creating new AWS Identity and Access Management (IAM) users. For those stacks, you must explicitly acknowledge this by specifying one of these capabilities.

The following IAM resources require you to specify either the CAPABILITY_IAM or CAPABILITY_NAMED_IAM capability.

If you have IAM resources, you can specify either capability.

If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

If you don't specify either of these capabilities, AWS CloudFormation returns an InsufficientCapabilities error.

If your stack template contains these resources, we recommend that you review all permissions associated with them and edit their permissions if necessary.

AWS::IAM::AccessKey

AWS::IAM::Group

AWS::IAM::InstanceProfile

AWS::IAM::Policy

AWS::IAM::Role

AWS::IAM::User

AWS::IAM::UserToGroupAddition

For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

CAPABILITY_AUTO_EXPAND

Some template contain macros. Macros perform custom processing on templates; this can include simple actions like find-and-replace operations, all the way to extensive transformations of entire templates. Because of this, users typically create a change set from the processed template, so that they can review the changes resulting from the macros before actually creating the stack. If your stack template contains one or more macros, and you choose to create a stack directly from the processed template, without first reviewing the resulting changes in a change set, you must acknowledge this capability. This includes the AWS::Include and AWS::Serverless transforms, which are macros hosted by AWS CloudFormation.

This capacity does not apply to creating change sets, and specifying it when creating change sets has no effect.

If you want to create a stack from a stack template that contains macros and nested stacks, you must create or update the stack directly from the template using the CreateStack or UpdateStack action, and specifying this capability.

For more information on macros, see Using AWS CloudFormation Macros to Perform Custom Processing on Templates.

pub fn set_capabilities(self, input: Option<Vec<Capability>>) -> Self

pub fn resource_types(self, inp: impl Into<String>) -> Self

The template resource types that you have permissions to work with if you execute this change set, such as AWS::EC2::Instance, AWS::EC2::*, or Custom::MyCustomInstance.

If the list of resource types doesn't include a resource type that you're updating, the stack update fails. By default, AWS CloudFormation grants permissions to all resource types. AWS Identity and Access Management (IAM) uses this parameter for condition keys in IAM policies for AWS CloudFormation. For more information, see Controlling Access with AWS Identity and Access Management in the AWS CloudFormation User Guide.

pub fn set_resource_types(self, input: Option<Vec<String>>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an AWS Identity and Access Management (IAM) role that AWS CloudFormation assumes when executing the change set. AWS CloudFormation uses the role's credentials to make calls on your behalf. AWS CloudFormation uses this role for all future operations on the stack. As long as users have permission to operate on the stack, AWS CloudFormation uses this role even if the users don't have permission to pass it. Ensure that the role grants least privilege.

If you don't specify a value, AWS CloudFormation uses the role that was previously associated with the stack. If no role is available, AWS CloudFormation uses a temporary session that is generated from your user credentials.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn rollback_configuration(self, input: RollbackConfiguration) -> Self

The rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

pub fn set_rollback_configuration(
self,
input: Option<RollbackConfiguration>
) -> Self

pub fn notification_ar_ns(self, inp: impl Into<String>) -> Self

The Amazon Resource Names (ARNs) of Amazon Simple Notification Service (Amazon SNS) topics that AWS CloudFormation associates with the stack. To remove all associated notification topics, specify an empty list.

pub fn set_notification_ar_ns(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Key-value pairs to associate with this stack. AWS CloudFormation also propagates these tags to resources in the stack. You can specify a maximum of 50 tags.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn change_set_name(self, input: impl Into<String>) -> Self

The name of the change set. The name must be unique among all change sets that are associated with the specified stack.

A change set name can contain only alphanumeric, case sensitive characters and hyphens. It must start with an alphabetic character and cannot exceed 128 characters.

pub fn set_change_set_name(self, input: Option<String>) -> Self

pub fn client_token(self, input: impl Into<String>) -> Self

A unique identifier for this CreateChangeSet request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to create another change set with the same name. You might retry CreateChangeSet requests to ensure that AWS CloudFormation successfully received them.

pub fn set_client_token(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description to help you identify this change set.

pub fn set_description(self, input: Option<String>) -> Self

pub fn change_set_type(self, input: ChangeSetType) -> Self

The type of change set operation. To create a change set for a new stack, specify CREATE. To create a change set for an existing stack, specify UPDATE. To create a change set for an import operation, specify IMPORT.

If you create a change set for a new stack, AWS Cloudformation creates a stack with a unique stack ID, but no template or resources. The stack will be in the REVIEW_IN_PROGRESS state until you execute the change set.

By default, AWS CloudFormation specifies UPDATE. You can't use the UPDATE type to create a change set for a new stack or the CREATE type to create a change set for an existing stack.

pub fn set_change_set_type(self, input: Option<ChangeSetType>) -> Self

pub fn resources_to_import(self, inp: impl Into<ResourceToImport>) -> Self

The resources to import into your stack.

pub fn set_resources_to_import(
self,
input: Option<Vec<ResourceToImport>>
) -> Self

pub fn include_nested_stacks(self, input: bool) -> Self

Creates a change set for the all nested stacks specified in the template. The default behavior of this action is set to False. To include nested sets in a change set, specify True.

pub fn set_include_nested_stacks(self, input: Option<bool>) -> Self