Struct aws_sdk_cloudformation::client::fluent_builders::CreateStackSet

pub struct CreateStackSet<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateStackSet<C>

pub async fn send(
self
) -> Result<CreateStackSetOutput, SdkError<CreateStackSetError>> where
C: SmithyConnector,

pub fn stack_set_name(self, input: impl Into<String>) -> Self

The name to associate with the stack set. The name must be unique in the Region where you create your stack set.

A stack name can contain only alphanumeric characters (case-sensitive) and hyphens. It must start with an alphabetic character and can't be longer than 128 characters.

pub fn set_stack_set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description of the stack set. You can use the description to identify the stack set's purpose or other important information.

pub fn set_description(self, input: Option<String>) -> Self

pub fn template_body(self, input: impl Into<String>) -> Self

The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

pub fn set_template_body(self, input: Option<String>) -> Self

pub fn template_url(self, input: impl Into<String>) -> Self

The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that's located in an Amazon S3 bucket or a Systems Manager document. For more information, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

pub fn set_template_url(self, input: Option<String>) -> Self

pub fn parameters(self, inp: impl Into<Parameter>) -> Self

The input parameters for the stack set template.

pub fn set_parameters(self, input: Option<Vec<Parameter>>) -> Self

pub fn capabilities(self, inp: impl Into<Capability>) -> Self

In some cases, you must explicitly acknowledge that your stack set template contains certain capabilities in order for AWS CloudFormation to create the stack set and related stack instances.

CAPABILITY_IAM and CAPABILITY_NAMED_IAM

Some stack templates might include resources that can affect permissions in your AWS account; for example, by creating new AWS Identity and Access Management (IAM) users. For those stack sets, you must explicitly acknowledge this by specifying one of these capabilities.

The following IAM resources require you to specify either the CAPABILITY_IAM or CAPABILITY_NAMED_IAM capability.

If you have IAM resources, you can specify either capability.

If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

If you don't specify either of these capabilities, AWS CloudFormation returns an InsufficientCapabilities error.

If your stack template contains these resources, we recommend that you review all permissions associated with them and edit their permissions if necessary.

AWS::IAM::AccessKey

AWS::IAM::Group

AWS::IAM::InstanceProfile

AWS::IAM::Policy

AWS::IAM::Role

AWS::IAM::User

AWS::IAM::UserToGroupAddition

For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

CAPABILITY_AUTO_EXPAND

Some templates reference macros. If your stack set template references one or more macros, you must create the stack set directly from the processed template, without first reviewing the resulting changes in a change set. To create the stack set directly, you must acknowledge this capability. For more information, see Using AWS CloudFormation Macros to Perform Custom Processing on Templates.

Stack sets with service-managed permissions do not currently support the use of macros in templates. (This includes the AWS::Include and AWS::Serverless transforms, which are macros hosted by AWS CloudFormation.) Even if you specify this capability for a stack set with service-managed permissions, if you reference a macro in your template the stack set operation will fail.

pub fn set_capabilities(self, input: Option<Vec<Capability>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

The key-value pairs to associate with this stack set and the stacks created from it. AWS CloudFormation also propagates these tags to supported resources that are created in the stacks. A maximum number of 50 tags can be specified.

If you specify tags as part of a CreateStackSet action, AWS CloudFormation checks to see if you have the required IAM permission to tag resources. If you don't, the entire CreateStackSet action fails with an access denied error, and the stack set is not created.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn administration_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Number (ARN) of the IAM role to use to create this stack set.

Specify an IAM role only if you are using customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Prerequisites: Granting Permissions for Stack Set Operations in the AWS CloudFormation User Guide.

pub fn set_administration_role_arn(self, input: Option<String>) -> Self

pub fn execution_role_name(self, input: impl Into<String>) -> Self

The name of the IAM execution role to use to create the stack set. If you do not specify an execution role, AWS CloudFormation uses the AWSCloudFormationStackSetExecutionRole role for the stack set operation.

Specify an IAM role only if you are using customized execution roles to control which stack resources users and groups can include in their stack sets.

pub fn set_execution_role_name(self, input: Option<String>) -> Self

pub fn permission_model(self, input: PermissionModels) -> Self

Describes how the IAM roles required for stack set operations are created. By default, SELF-MANAGED is specified.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by AWS Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

pub fn set_permission_model(self, input: Option<PermissionModels>) -> Self

pub fn auto_deployment(self, input: AutoDeployment) -> Self

Describes whether StackSets automatically deploys to AWS Organizations accounts that are added to the target organization or organizational unit (OU). Specify only if PermissionModel is SERVICE_MANAGED.

pub fn set_auto_deployment(self, input: Option<AutoDeployment>) -> Self

pub fn call_as(self, input: CallAs) -> Self

[Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

To create a stack set with service-managed permissions while signed in to the management account, specify SELF.

To create a stack set with service-managed permissions while signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your AWS account must be registered as a delegated admin in the management account. For more information, see Register a delegated administrator in the AWS CloudFormation User Guide.

Stack sets with service-managed permissions are created in the management account, including stack sets that are created by delegated administrators.

pub fn set_call_as(self, input: Option<CallAs>) -> Self

pub fn client_request_token(self, input: impl Into<String>) -> Self

A unique identifier for this CreateStackSet request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to create another stack set with the same name. You might retry CreateStackSet requests to ensure that AWS CloudFormation successfully received them.

If you don't specify an operation ID, the SDK generates one automatically.

pub fn set_client_request_token(self, input: Option<String>) -> Self