Struct aws_sdk_cloudformation::client::fluent_builders::UpdateStack

pub struct UpdateStack<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> UpdateStack<C>

pub async fn send(self) -> Result<UpdateStackOutput, SdkError<UpdateStackError>> where
C: SmithyConnector,

pub fn stack_name(self, input: impl Into<String>) -> Self

The name or unique stack ID of the stack to update.

pub fn set_stack_name(self, input: Option<String>) -> Self

pub fn template_body(self, input: impl Into<String>) -> Self

Structure containing the template body with a minimum length of 1 byte and a maximum length of 51,200 bytes. (For more information, go to Template Anatomy in the AWS CloudFormation User Guide.)

Conditional: You must specify only one of the following parameters: TemplateBody, TemplateURL, or set the UsePreviousTemplate to true.

pub fn set_template_body(self, input: Option<String>) -> Self

pub fn template_url(self, input: impl Into<String>) -> Self

Location of file containing the template body. The URL must point to a template that is located in an Amazon S3 bucket or a Systems Manager document. For more information, go to Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: TemplateBody, TemplateURL, or set the UsePreviousTemplate to true.

pub fn set_template_url(self, input: Option<String>) -> Self

pub fn use_previous_template(self, input: bool) -> Self

Reuse the existing template that is associated with the stack that you are updating.

Conditional: You must specify only one of the following parameters: TemplateBody, TemplateURL, or set the UsePreviousTemplate to true.

pub fn set_use_previous_template(self, input: Option<bool>) -> Self

pub fn stack_policy_during_update_body(self, input: impl Into<String>) -> Self

Structure containing the temporary overriding stack policy body. You can specify either the StackPolicyDuringUpdateBody or the StackPolicyDuringUpdateURL parameter, but not both.

If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.

pub fn set_stack_policy_during_update_body(self, input: Option<String>) -> Self

pub fn stack_policy_during_update_url(self, input: impl Into<String>) -> Self

Location of a file containing the temporary overriding stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same Region as the stack. You can specify either the StackPolicyDuringUpdateBody or the StackPolicyDuringUpdateURL parameter, but not both.

If you want to update protected resources, specify a temporary overriding stack policy during this update. If you do not specify a stack policy, the current policy that is associated with the stack will be used.

pub fn set_stack_policy_during_update_url(self, input: Option<String>) -> Self

pub fn parameters(self, inp: impl Into<Parameter>) -> Self

A list of Parameter structures that specify input parameters for the stack. For more information, see the Parameter data type.

pub fn set_parameters(self, input: Option<Vec<Parameter>>) -> Self

pub fn capabilities(self, inp: impl Into<Capability>) -> Self

In some cases, you must explicitly acknowledge that your stack template contains certain capabilities in order for AWS CloudFormation to update the stack.

CAPABILITY_IAM and CAPABILITY_NAMED_IAM

Some stack templates might include resources that can affect permissions in your AWS account; for example, by creating new AWS Identity and Access Management (IAM) users. For those stacks, you must explicitly acknowledge this by specifying one of these capabilities.

The following IAM resources require you to specify either the CAPABILITY_IAM or CAPABILITY_NAMED_IAM capability.

If you have IAM resources, you can specify either capability.

If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

If you don't specify either of these capabilities, AWS CloudFormation returns an InsufficientCapabilities error.

If your stack template contains these resources, we recommend that you review all permissions associated with them and edit their permissions if necessary.

AWS::IAM::AccessKey

AWS::IAM::Group

AWS::IAM::InstanceProfile

AWS::IAM::Policy

AWS::IAM::Role

AWS::IAM::User

AWS::IAM::UserToGroupAddition

For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

CAPABILITY_AUTO_EXPAND

Some template contain macros. Macros perform custom processing on templates; this can include simple actions like find-and-replace operations, all the way to extensive transformations of entire templates. Because of this, users typically create a change set from the processed template, so that they can review the changes resulting from the macros before actually updating the stack. If your stack template contains one or more macros, and you choose to update a stack directly from the processed template, without first reviewing the resulting changes in a change set, you must acknowledge this capability. This includes the AWS::Include and AWS::Serverless transforms, which are macros hosted by AWS CloudFormation.

If you want to update a stack from a stack template that contains macros and nested stacks, you must update the stack directly from the template using this capability.

You should only update stacks directly from a stack template that contains macros if you know what processing the macro performs.

Each macro relies on an underlying Lambda service function for processing stack templates. Be aware that the Lambda function owner can update the function operation without AWS CloudFormation being notified.

For more information, see Using AWS CloudFormation Macros to Perform Custom Processing on Templates.

pub fn set_capabilities(self, input: Option<Vec<Capability>>) -> Self

pub fn resource_types(self, inp: impl Into<String>) -> Self

The template resource types that you have permissions to work with for this update stack action, such as AWS::EC2::Instance, AWS::EC2::*, or Custom::MyCustomInstance.

If the list of resource types doesn't include a resource that you're updating, the stack update fails. By default, AWS CloudFormation grants permissions to all resource types. AWS Identity and Access Management (IAM) uses this parameter for AWS CloudFormation-specific condition keys in IAM policies. For more information, see Controlling Access with AWS Identity and Access Management.

pub fn set_resource_types(self, input: Option<Vec<String>>) -> Self

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of an AWS Identity and Access Management (IAM) role that AWS CloudFormation assumes to update the stack. AWS CloudFormation uses the role's credentials to make calls on your behalf. AWS CloudFormation always uses this role for all future operations on the stack. As long as users have permission to operate on the stack, AWS CloudFormation uses this role even if the users don't have permission to pass it. Ensure that the role grants least privilege.

If you don't specify a value, AWS CloudFormation uses the role that was previously associated with the stack. If no role is available, AWS CloudFormation uses a temporary session that is generated from your user credentials.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn rollback_configuration(self, input: RollbackConfiguration) -> Self

The rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

pub fn set_rollback_configuration(
self,
input: Option<RollbackConfiguration>
) -> Self

pub fn stack_policy_body(self, input: impl Into<String>) -> Self

Structure containing a new stack policy body. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.

pub fn set_stack_policy_body(self, input: Option<String>) -> Self

pub fn stack_policy_url(self, input: impl Into<String>) -> Self

Location of a file containing the updated stack policy. The URL must point to a policy (max size: 16KB) located in an S3 bucket in the same Region as the stack. You can specify either the StackPolicyBody or the StackPolicyURL parameter, but not both.

You might update the stack policy, for example, in order to protect a new resource that you created during a stack update. If you do not specify a stack policy, the current policy that is associated with the stack is unchanged.

pub fn set_stack_policy_url(self, input: Option<String>) -> Self

pub fn notification_ar_ns(self, inp: impl Into<String>) -> Self

Amazon Simple Notification Service topic Amazon Resource Names (ARNs) that AWS CloudFormation associates with the stack. Specify an empty list to remove all notification topics.

pub fn set_notification_ar_ns(self, input: Option<Vec<String>>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Key-value pairs to associate with this stack. AWS CloudFormation also propagates these tags to supported resources in the stack. You can specify a maximum number of 50 tags.

If you don't specify this parameter, AWS CloudFormation doesn't modify the stack's tags. If you specify an empty value, AWS CloudFormation removes all associated tags.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn client_request_token(self, input: impl Into<String>) -> Self

A unique identifier for this UpdateStack request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to update a stack with the same name. You might retry UpdateStack requests to ensure that AWS CloudFormation successfully received them.

All events triggered by a given stack operation are assigned the same client request token, which you can use to track operations. For example, if you execute a CreateStack operation with the token token1, then all the StackEvents generated by that operation will have ClientRequestToken set as token1.

In the console, stack operations display the client request token on the Events tab. Stack operations that are initiated from the console use the token format Console-StackOperation-ID, which helps you easily identify the stack operation . For example, if you create a stack using the console, each stack event would be assigned the same token in the following format: Console-CreateStack-7f59c3cf-00d2-40c7-b2ff-e75db0987002.

pub fn set_client_request_token(self, input: Option<String>) -> Self