Module aws_sdk_cloudformation::error

Modules

already_exists_exception

See AlreadyExistsException

cfn_registry_exception

See CfnRegistryException

change_set_not_found_exception

See ChangeSetNotFoundException

created_but_modified_exception

See CreatedButModifiedException

insufficient_capabilities_exception

See InsufficientCapabilitiesException

invalid_change_set_status_exception

See InvalidChangeSetStatusException

invalid_operation_exception

See InvalidOperationException

invalid_state_transition_exception

See InvalidStateTransitionException

limit_exceeded_exception

See LimitExceededException

name_already_exists_exception

See NameAlreadyExistsException

operation_id_already_exists_exception

See OperationIdAlreadyExistsException

operation_in_progress_exception

See OperationInProgressException

operation_not_found_exception

See OperationNotFoundException

operation_status_check_failed_exception

See OperationStatusCheckFailedException

stack_instance_not_found_exception

See StackInstanceNotFoundException

stack_set_not_empty_exception

See StackSetNotEmptyException

stack_set_not_found_exception

See StackSetNotFoundException

stale_request_exception

See StaleRequestException

token_already_exists_exception

See TokenAlreadyExistsException

type_configuration_not_found_exception

See TypeConfigurationNotFoundException

type_not_found_exception

See TypeNotFoundException

Structs

ActivateTypeError

AlreadyExistsException

The resource with the name requested already exists.

BatchDescribeTypeConfigurationsError

CancelUpdateStackError

CfnRegistryException

An error occurred during a CloudFormation registry operation.

ChangeSetNotFoundException

The specified change set name or ID doesn't exit. To view valid change sets for a

ContinueUpdateRollbackError

CreateChangeSetError

CreateStackError

CreateStackInstancesError

CreateStackSetError

CreatedButModifiedException

The specified resource exists, but has been changed.

DeactivateTypeError

DeleteChangeSetError

DeleteStackError

DeleteStackInstancesError

DeleteStackSetError

DeregisterTypeError

DescribeAccountLimitsError

DescribeChangeSetError

DescribePublisherError

DescribeStackDriftDetectionStatusError

DescribeStackEventsError

DescribeStackInstanceError

DescribeStackResourceDriftsError

DescribeStackResourceError

DescribeStackResourcesError

DescribeStackSetError

DescribeStackSetOperationError

DescribeStacksError

DescribeTypeError

DescribeTypeRegistrationError

DetectStackDriftError

DetectStackResourceDriftError

DetectStackSetDriftError

EstimateTemplateCostError

ExecuteChangeSetError

GetStackPolicyError

GetTemplateError

GetTemplateSummaryError

InsufficientCapabilitiesException

The template contains resources with capabilities that weren't specified in the

InvalidChangeSetStatusException

The specified change set can't be used to update the stack. For example, the change

InvalidOperationException

The specified operation isn't valid.

InvalidStateTransitionException

Error reserved for use by the CloudFormation CLI. CloudFormation does not return this error to users.

LimitExceededException

The quota for the resource has already been reached.

ListChangeSetsError

ListExportsError

ListImportsError

ListStackInstancesError

ListStackResourcesError

ListStackSetOperationResultsError

ListStackSetOperationsError

ListStackSetsError

ListStacksError

ListTypeRegistrationsError

ListTypeVersionsError

ListTypesError

NameAlreadyExistsException

The specified name is already in use.

OperationIdAlreadyExistsException

The specified operation ID already exists.

OperationInProgressException

Another operation is currently in progress for this stack set. Only one operation can

OperationNotFoundException

The specified ID refers to an operation that doesn't exist.

OperationStatusCheckFailedException

Error reserved for use by the CloudFormation CLI. CloudFormation does not return this error to users.

PublishTypeError

RecordHandlerProgressError

RegisterPublisherError

RegisterTypeError

SetStackPolicyError

SetTypeConfigurationError

SetTypeDefaultVersionError

SignalResourceError

StackInstanceNotFoundException

The specified stack instance doesn't exist.

StackSetNotEmptyException

You can't yet delete this stack set, because it still contains one or more stack

StackSetNotFoundException

The specified stack set doesn't exist.

StaleRequestException

Another operation has been performed on this stack set since the specified operation

StopStackSetOperationError

TestTypeError

TokenAlreadyExistsException

A client request token already exists.

TypeConfigurationNotFoundException

The specified extension configuration cannot be found.

TypeNotFoundException

The specified extension does not exist in the CloudFormation registry.

UpdateStackError

UpdateStackInstancesError

UpdateStackSetError

UpdateTerminationProtectionError

ValidateTemplateError

Enums

ActivateTypeErrorKind

BatchDescribeTypeConfigurationsErrorKind

CancelUpdateStackErrorKind

ContinueUpdateRollbackErrorKind

CreateChangeSetErrorKind

CreateStackErrorKind

CreateStackInstancesErrorKind

CreateStackSetErrorKind

DeactivateTypeErrorKind

DeleteChangeSetErrorKind

DeleteStackErrorKind

DeleteStackInstancesErrorKind

DeleteStackSetErrorKind

DeregisterTypeErrorKind

DescribeAccountLimitsErrorKind

DescribeChangeSetErrorKind

DescribePublisherErrorKind

DescribeStackDriftDetectionStatusErrorKind

DescribeStackEventsErrorKind

DescribeStackInstanceErrorKind

DescribeStackResourceDriftsErrorKind

DescribeStackResourceErrorKind

DescribeStackResourcesErrorKind

DescribeStackSetErrorKind

DescribeStackSetOperationErrorKind

DescribeStacksErrorKind

DescribeTypeErrorKind

DescribeTypeRegistrationErrorKind

DetectStackDriftErrorKind

DetectStackResourceDriftErrorKind

DetectStackSetDriftErrorKind

EstimateTemplateCostErrorKind

ExecuteChangeSetErrorKind

GetStackPolicyErrorKind

GetTemplateErrorKind

GetTemplateSummaryErrorKind

ListChangeSetsErrorKind

ListExportsErrorKind

ListImportsErrorKind

ListStackInstancesErrorKind

ListStackResourcesErrorKind

ListStackSetOperationResultsErrorKind

ListStackSetOperationsErrorKind

ListStackSetsErrorKind

ListStacksErrorKind

ListTypeRegistrationsErrorKind

ListTypeVersionsErrorKind

ListTypesErrorKind

PublishTypeErrorKind

RecordHandlerProgressErrorKind

RegisterPublisherErrorKind

RegisterTypeErrorKind

SetStackPolicyErrorKind

SetTypeConfigurationErrorKind

SetTypeDefaultVersionErrorKind

SignalResourceErrorKind

StopStackSetOperationErrorKind

TestTypeErrorKind

UpdateStackErrorKind

UpdateStackInstancesErrorKind

UpdateStackSetErrorKind

UpdateTerminationProtectionErrorKind

ValidateTemplateErrorKind