Struct aws_sdk_cloudformation::input::CreateStackSetInput

#[non_exhaustive]
pub struct CreateStackSetInput {

    pub stack_set_name: Option<String>,
    pub description: Option<String>,
    pub template_body: Option<String>,
    pub template_url: Option<String>,
    pub parameters: Option<Vec<Parameter>>,
    pub capabilities: Option<Vec<Capability>>,
    pub tags: Option<Vec<Tag>>,
    pub administration_role_arn: Option<String>,
    pub execution_role_name: Option<String>,
    pub permission_model: Option<PermissionModels>,
    pub auto_deployment: Option<AutoDeployment>,
    pub call_as: Option<CallAs>,
    pub client_request_token: Option<String>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

stack_set_name: Option<String>

The name to associate with the stack set. The name must be unique in the Region where you create your stack set.

A stack name can contain only alphanumeric characters (case-sensitive) and hyphens. It must start with an alphabetic character and can't be longer than 128 characters.

description: Option<String>

A description of the stack set. You can use the description to identify the stack set's purpose or other important information.

template_body: Option<String>

The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

template_url: Option<String>

The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that's located in an Amazon S3 bucket or a Systems Manager document. For more information, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify either the TemplateBody or the TemplateURL parameter, but not both.

parameters: Option<Vec<Parameter>>

The input parameters for the stack set template.

capabilities: Option<Vec<Capability>>

In some cases, you must explicitly acknowledge that your stack set template contains certain capabilities in order for AWS CloudFormation to create the stack set and related stack instances.

CAPABILITY_IAM and CAPABILITY_NAMED_IAM

Some stack templates might include resources that can affect permissions in your AWS account; for example, by creating new AWS Identity and Access Management (IAM) users. For those stack sets, you must explicitly acknowledge this by specifying one of these capabilities.

The following IAM resources require you to specify either the CAPABILITY_IAM or CAPABILITY_NAMED_IAM capability.

If you have IAM resources, you can specify either capability.

If you have IAM resources with custom names, you must specify CAPABILITY_NAMED_IAM.

If you don't specify either of these capabilities, AWS CloudFormation returns an InsufficientCapabilities error.

If your stack template contains these resources, we recommend that you review all permissions associated with them and edit their permissions if necessary.

AWS::IAM::AccessKey

AWS::IAM::Group

AWS::IAM::InstanceProfile

AWS::IAM::Policy

AWS::IAM::Role

AWS::IAM::User

AWS::IAM::UserToGroupAddition

For more information, see Acknowledging IAM Resources in AWS CloudFormation Templates.

CAPABILITY_AUTO_EXPAND

Some templates reference macros. If your stack set template references one or more macros, you must create the stack set directly from the processed template, without first reviewing the resulting changes in a change set. To create the stack set directly, you must acknowledge this capability. For more information, see Using AWS CloudFormation Macros to Perform Custom Processing on Templates.

Stack sets with service-managed permissions do not currently support the use of macros in templates. (This includes the AWS::Include and AWS::Serverless transforms, which are macros hosted by AWS CloudFormation.) Even if you specify this capability for a stack set with service-managed permissions, if you reference a macro in your template the stack set operation will fail.

tags: Option<Vec<Tag>>

The key-value pairs to associate with this stack set and the stacks created from it. AWS CloudFormation also propagates these tags to supported resources that are created in the stacks. A maximum number of 50 tags can be specified.

If you specify tags as part of a CreateStackSet action, AWS CloudFormation checks to see if you have the required IAM permission to tag resources. If you don't, the entire CreateStackSet action fails with an access denied error, and the stack set is not created.

administration_role_arn: Option<String>

The Amazon Resource Number (ARN) of the IAM role to use to create this stack set.

Specify an IAM role only if you are using customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Prerequisites: Granting Permissions for Stack Set Operations in the AWS CloudFormation User Guide.

execution_role_name: Option<String>

The name of the IAM execution role to use to create the stack set. If you do not specify an execution role, AWS CloudFormation uses the AWSCloudFormationStackSetExecutionRole role for the stack set operation.

Specify an IAM role only if you are using customized execution roles to control which stack resources users and groups can include in their stack sets.

permission_model: Option<PermissionModels>

Describes how the IAM roles required for stack set operations are created. By default, SELF-MANAGED is specified.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by AWS Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

auto_deployment: Option<AutoDeployment>

Describes whether StackSets automatically deploys to AWS Organizations accounts that are added to the target organization or organizational unit (OU). Specify only if PermissionModel is SERVICE_MANAGED.

call_as: Option<CallAs>

[Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

To create a stack set with service-managed permissions while signed in to the management account, specify SELF.

To create a stack set with service-managed permissions while signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your AWS account must be registered as a delegated admin in the management account. For more information, see Register a delegated administrator in the AWS CloudFormation User Guide.

Stack sets with service-managed permissions are created in the management account, including stack sets that are created by delegated administrators.

client_request_token: Option<String>

A unique identifier for this CreateStackSet request. Specify this token if you plan to retry requests so that AWS CloudFormation knows that you're not attempting to create another stack set with the same name. You might retry CreateStackSet requests to ensure that AWS CloudFormation successfully received them.

If you don't specify an operation ID, the SDK generates one automatically.

Implementations

impl CreateStackSetInput

pub fn make_operation(
self,
_config: &Config
) -> Result<Operation<CreateStackSet, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateStackSet>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateStackSetInput