Struct aws_sdk_cloudformation::input::update_stack_set_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UpdateStackSetInput

Implementations

impl Builder

pub fn stack_set_name(self, input: impl Into<String>) -> Self

The name or unique ID of the stack set that you want to update.

pub fn set_stack_set_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A brief description of updates that you are making.

pub fn set_description(self, input: Option<String>) -> Self

pub fn template_body(self, input: impl Into<String>) -> Self

The structure that contains the template body, with a minimum length of 1 byte and a maximum length of 51,200 bytes. For more information, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL—or set UsePreviousTemplate to true.

pub fn set_template_body(self, input: Option<String>) -> Self

pub fn template_url(self, input: impl Into<String>) -> Self

The location of the file that contains the template body. The URL must point to a template (maximum size: 460,800 bytes) that is located in an Amazon S3 bucket or a Systems Manager document. For more information, see Template Anatomy in the AWS CloudFormation User Guide.

Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL—or set UsePreviousTemplate to true.

pub fn set_template_url(self, input: Option<String>) -> Self

pub fn use_previous_template(self, input: bool) -> Self

Use the existing template that's associated with the stack set that you're updating.

Conditional: You must specify only one of the following parameters: TemplateBody or TemplateURL—or set UsePreviousTemplate to true.

pub fn set_use_previous_template(self, input: Option<bool>) -> Self

pub fn parameters(self, input: impl Into<Parameter>) -> Self

pub fn set_parameters(self, input: Option<Vec<Parameter>>) -> Self

pub fn capabilities(self, input: impl Into<Capability>) -> Self

pub fn set_capabilities(self, input: Option<Vec<Capability>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn operation_preferences(self, input: StackSetOperationPreferences) -> Self

Preferences for how AWS CloudFormation performs this stack set operation.

pub fn set_operation_preferences(
self,
input: Option<StackSetOperationPreferences>
) -> Self

pub fn administration_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Number (ARN) of the IAM role to use to update this stack set.

Specify an IAM role only if you are using customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Granting Permissions for Stack Set Operations in the AWS CloudFormation User Guide.

If you specified a customized administrator role when you created the stack set, you must specify a customized administrator role, even if it is the same customized administrator role used with this stack set previously.

pub fn set_administration_role_arn(self, input: Option<String>) -> Self

pub fn execution_role_name(self, input: impl Into<String>) -> Self

The name of the IAM execution role to use to update the stack set. If you do not specify an execution role, AWS CloudFormation uses the AWSCloudFormationStackSetExecutionRole role for the stack set operation.

Specify an IAM role only if you are using customized execution roles to control which stack resources users and groups can include in their stack sets.

If you specify a customized execution role, AWS CloudFormation uses that role to update the stack. If you do not specify a customized execution role, AWS CloudFormation performs the update using the role previously associated with the stack set, so long as you have permissions to perform operations on the stack set.

pub fn set_execution_role_name(self, input: Option<String>) -> Self

pub fn deployment_targets(self, input: DeploymentTargets) -> Self

[Service-managed permissions] The AWS Organizations accounts in which to update associated stack instances.

To update all the stack instances associated with this stack set, do not specify DeploymentTargets or Regions.

If the stack set update includes changes to the template (that is, if TemplateBody or TemplateURL is specified), or the Parameters, AWS CloudFormation marks all stack instances with a status of OUTDATED prior to updating the stack instances in the specified accounts and Regions. If the stack set update does not include changes to the template or parameters, AWS CloudFormation updates the stack instances in the specified accounts and Regions, while leaving all other stack instances with their existing stack instance status.

pub fn set_deployment_targets(self, input: Option<DeploymentTargets>) -> Self

pub fn permission_model(self, input: PermissionModels) -> Self

Describes how the IAM roles required for stack set operations are created. You cannot modify PermissionModel if there are stack instances associated with your stack set.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by AWS Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

pub fn set_permission_model(self, input: Option<PermissionModels>) -> Self

pub fn auto_deployment(self, input: AutoDeployment) -> Self

[Service-managed permissions] Describes whether StackSets automatically deploys to AWS Organizations accounts that are added to a target organization or organizational unit (OU).

If you specify AutoDeployment, do not specify DeploymentTargets or Regions.

pub fn set_auto_deployment(self, input: Option<AutoDeployment>) -> Self

pub fn operation_id(self, input: impl Into<String>) -> Self

The unique ID for this stack set operation.

The operation ID also functions as an idempotency token, to ensure that AWS CloudFormation performs the stack set operation only once, even if you retry the request multiple times. You might retry stack set operation requests to ensure that AWS CloudFormation successfully received them.

If you don't specify an operation ID, AWS CloudFormation generates one automatically.

Repeating this stack set operation with a new operation ID retries all stack instances whose status is OUTDATED.

pub fn set_operation_id(self, input: Option<String>) -> Self

pub fn accounts(self, input: impl Into<String>) -> Self

pub fn set_accounts(self, input: Option<Vec<String>>) -> Self

pub fn regions(self, input: impl Into<String>) -> Self

pub fn set_regions(self, input: Option<Vec<String>>) -> Self

pub fn call_as(self, input: CallAs) -> Self

[Service-managed permissions] Specifies whether you are acting as an account administrator in the organization's management account or as a delegated administrator in a member account.

By default, SELF is specified. Use SELF for stack sets with self-managed permissions.

If you are signed in to the management account, specify SELF.

If you are signed in to a delegated administrator account, specify DELEGATED_ADMIN.

Your AWS account must be registered as a delegated administrator in the management account. For more information, see Register a delegated administrator in the AWS CloudFormation User Guide.

pub fn set_call_as(self, input: Option<CallAs>) -> Self

pub fn build(self) -> Result<UpdateStackSetInput, BuildError>

Consumes the builder and constructs a UpdateStackSetInput