Module aws_sdk_cloudformation::model

Modules

account_gate_result

See AccountGateResult

account_limit

See AccountLimit

auto_deployment

See AutoDeployment

batch_describe_type_configurations_error

See BatchDescribeTypeConfigurationsError

change

See Change

change_set_summary

See ChangeSetSummary

deployment_targets

See DeploymentTargets

export

See Export

logging_config

See LoggingConfig

module_info

See ModuleInfo

output

See Output

parameter

See Parameter

parameter_constraints

See ParameterConstraints

parameter_declaration

See ParameterDeclaration

physical_resource_id_context_key_value_pair

See PhysicalResourceIdContextKeyValuePair

property_difference

See PropertyDifference

required_activated_type

See RequiredActivatedType

resource_change

See ResourceChange

resource_change_detail

See ResourceChangeDetail

resource_identifier_summary

See ResourceIdentifierSummary

resource_target_definition

See ResourceTargetDefinition

resource_to_import

See ResourceToImport

rollback_configuration

See RollbackConfiguration

rollback_trigger

See RollbackTrigger

stack

See Stack

stack_drift_information

See StackDriftInformation

stack_drift_information_summary

See StackDriftInformationSummary

stack_event

See StackEvent

stack_instance

See StackInstance

stack_instance_comprehensive_status

See StackInstanceComprehensiveStatus

stack_instance_filter

See StackInstanceFilter

stack_instance_summary

See StackInstanceSummary

stack_resource

See StackResource

stack_resource_detail

See StackResourceDetail

stack_resource_drift

See StackResourceDrift

stack_resource_drift_information

See StackResourceDriftInformation

stack_resource_drift_information_summary

See StackResourceDriftInformationSummary

stack_resource_summary

See StackResourceSummary

stack_set

See StackSet

stack_set_drift_detection_details

See StackSetDriftDetectionDetails

stack_set_operation

See StackSetOperation

stack_set_operation_preferences

See StackSetOperationPreferences

stack_set_operation_result_summary

See StackSetOperationResultSummary

stack_set_operation_summary

See StackSetOperationSummary

stack_set_summary

See StackSetSummary

stack_summary

See StackSummary

tag

See Tag

template_parameter

See TemplateParameter

type_configuration_details

See TypeConfigurationDetails

type_configuration_identifier

See TypeConfigurationIdentifier

type_filters

See TypeFilters

type_summary

See TypeSummary

type_version_summary

See TypeVersionSummary

Structs

AccountGateResult

Structure that contains the results of the account gate function which AWS

AccountLimit

The AccountLimit data type.

AutoDeployment

[Service-managed permissions] Describes whether StackSets automatically deploys to AWS Organizations accounts that are added to a target organization or organizational unit (OU).

BatchDescribeTypeConfigurationsError

Detailed information concerning an error generated during the setting of configuration data for a CloudFormation extension.

Change

The Change structure describes the changes AWS CloudFormation will

ChangeSetSummary

The ChangeSetSummary structure describes a change set, its status, and

DeploymentTargets

[Service-managed permissions] The AWS Organizations accounts to which StackSets deploys. StackSets does not deploy stack instances to the organization management account, even if the organization management account is in your organization or in an OU in your organization.

Export

The Export structure describes the exported output values for a

LoggingConfig

Contains logging configuration information for an extension.

ModuleInfo

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

Output

The Output data type.

Parameter

The Parameter data type.

ParameterConstraints

A set of criteria that AWS CloudFormation uses to validate parameter values. Although

ParameterDeclaration

The ParameterDeclaration data type.

PhysicalResourceIdContextKeyValuePair

Context information that enables AWS CloudFormation to uniquely identify a resource. AWS CloudFormation uses

PropertyDifference

Information about a resource property whose actual value differs from its expected

RequiredActivatedType

For extensions that are modules, a public third-party extension that must be activated in your account in order for the module itself to be activated.

ResourceChange

The ResourceChange structure describes the resource and the action that

ResourceChangeDetail

For a resource with Modify as the action, the

ResourceIdentifierSummary

Describes the target resources of a specific type in your import template (for example,

ResourceTargetDefinition

The field that AWS CloudFormation will change, such as the name of a resource's

ResourceToImport

Describes the target resource of an import operation.

RollbackConfiguration

Structure containing the rollback triggers for AWS CloudFormation to monitor during stack creation

RollbackTrigger

A rollback trigger AWS CloudFormation monitors during creation and updating of stacks. If any of

Stack

The Stack data type.

StackDriftInformation

Contains information about whether the stack's actual configuration differs, or has

StackDriftInformationSummary

Contains information about whether the stack's actual configuration differs, or has

StackEvent

The StackEvent data type.

StackInstance

An AWS CloudFormation stack, in a specific account and Region, that's part of a stack

StackInstanceComprehensiveStatus

The detailed status of the stack instance.

StackInstanceFilter

The status that stack instances are filtered by.

StackInstanceSummary

The structure that contains summary information about a stack instance.

StackResource

The StackResource data type.

StackResourceDetail

Contains detailed information about the specified stack resource.

StackResourceDrift

Contains the drift information for a resource that has been checked for drift. This

StackResourceDriftInformation

Contains information about whether the resource's actual configuration differs, or

StackResourceDriftInformationSummary

Summarizes information about whether the resource's actual configuration differs, or

StackResourceSummary

Contains high-level information about the specified stack resource.

StackSet

A structure that contains information about a stack set. A stack set enables you to

StackSetDriftDetectionDetails

Detailed information about the drift status of the stack set.

StackSetOperation

The structure that contains information about a stack set operation.

StackSetOperationPreferences

The user-specified preferences for how AWS CloudFormation performs a stack set

StackSetOperationResultSummary

The structure that contains information about a specified operation's results for a

StackSetOperationSummary

The structures that contain summary information about the specified

StackSetSummary

The structures that contain summary information about the specified stack

StackSummary

The StackSummary Data Type

Tag

The Tag type enables you to specify a key-value pair that can be used to store

TemplateParameter

The TemplateParameter data type.

TypeConfigurationDetails

Detailed information concerning the specification of a CloudFormation extension in a given account and region.

TypeConfigurationIdentifier

Identifying information for the configuration of a CloudFormation extension.

TypeFilters

Filter criteria to use in determining which extensions to return.

TypeSummary

Contains summary information about the specified CloudFormation extension.

TypeVersionSummary

Contains summary information about a specific version of a CloudFormation extension.

Enums

AccountGateStatus

CallAs

NOTE: CallAs::Self has been renamed to ::SelfValue.

Capability

Category

ChangeAction

ChangeSetStatus

ChangeSetType

ChangeSource

ChangeType

DeprecatedStatus

DifferenceType

EvaluationType

ExecutionStatus

HandlerErrorCode

IdentityProvider

OnFailure

OperationStatus

PermissionModels

ProvisioningType

PublisherStatus

RegionConcurrencyType

RegistrationStatus

RegistryType

Replacement

RequiresRecreation

ResourceAttribute

ResourceSignalStatus

ResourceStatus

StackDriftDetectionStatus

StackDriftStatus

NOTE: StackDriftStatus::Unknown has been renamed to ::UnknownValue.

StackInstanceDetailedStatus

StackInstanceFilterName

StackInstanceStatus

StackResourceDriftStatus

StackSetDriftDetectionStatus

StackSetDriftStatus

StackSetOperationAction

StackSetOperationResultStatus

StackSetOperationStatus

StackSetStatus

StackStatus

TemplateStage

ThirdPartyType

TypeTestsStatus

VersionBump

Visibility