Struct aws_sdk_cloudformation::model::stack_resource_drift::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for StackResourceDrift

Implementations

impl Builder

pub fn stack_id(self, input: impl Into<String>) -> Self

The ID of the stack.

pub fn set_stack_id(self, input: Option<String>) -> Self

pub fn logical_resource_id(self, input: impl Into<String>) -> Self

The logical name of the resource specified in the template.

pub fn set_logical_resource_id(self, input: Option<String>) -> Self

pub fn physical_resource_id(self, input: impl Into<String>) -> Self

The name or unique identifier that corresponds to a physical instance ID of a resource supported by AWS CloudFormation.

pub fn set_physical_resource_id(self, input: Option<String>) -> Self

pub fn physical_resource_id_context(
self,
input: impl Into<PhysicalResourceIdContextKeyValuePair>
) -> Self

pub fn set_physical_resource_id_context(
self,
input: Option<Vec<PhysicalResourceIdContextKeyValuePair>>
) -> Self

pub fn resource_type(self, input: impl Into<String>) -> Self

The type of the resource.

pub fn set_resource_type(self, input: Option<String>) -> Self

pub fn expected_properties(self, input: impl Into<String>) -> Self

A JSON structure containing the expected property values of the stack resource, as defined in the stack template and any values specified as template parameters.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

pub fn set_expected_properties(self, input: Option<String>) -> Self

pub fn actual_properties(self, input: impl Into<String>) -> Self

A JSON structure containing the actual property values of the stack resource.

For resources whose StackResourceDriftStatus is DELETED, this structure will not be present.

pub fn set_actual_properties(self, input: Option<String>) -> Self

pub fn property_differences(self, input: impl Into<PropertyDifference>) -> Self

pub fn set_property_differences(
self,
input: Option<Vec<PropertyDifference>>
) -> Self

pub fn stack_resource_drift_status(
self,
input: StackResourceDriftStatus
) -> Self

Status of the resource's actual configuration compared to its expected configuration

DELETED: The resource differs from its expected template configuration because the resource has been deleted.

MODIFIED: One or more resource properties differ from their expected values (as defined in the stack template and any values specified as template parameters).

IN_SYNC: The resources's actual configuration matches its expected template configuration.

NOT_CHECKED: AWS CloudFormation does not currently return this value.

pub fn set_stack_resource_drift_status(
self,
input: Option<StackResourceDriftStatus>
) -> Self

pub fn timestamp(self, input: Instant) -> Self

Time at which AWS CloudFormation performed drift detection on the stack resource.

pub fn set_timestamp(self, input: Option<Instant>) -> Self

pub fn module_info(self, input: ModuleInfo) -> Self

Contains information about the module from which the resource was created, if the resource was created from a module included in the stack template.

pub fn set_module_info(self, input: Option<ModuleInfo>) -> Self

pub fn build(self) -> StackResourceDrift

Consumes the builder and constructs a StackResourceDrift