Struct aws_sdk_cloudformation::model::stack_set::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for StackSet

Implementations

impl Builder

pub fn stack_set_name(self, input: impl Into<String>) -> Self

The name that's associated with the stack set.

pub fn set_stack_set_name(self, input: Option<String>) -> Self

pub fn stack_set_id(self, input: impl Into<String>) -> Self

The ID of the stack set.

pub fn set_stack_set_id(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description of the stack set that you specify when the stack set is created or updated.

pub fn set_description(self, input: Option<String>) -> Self

pub fn status(self, input: StackSetStatus) -> Self

The status of the stack set.

pub fn set_status(self, input: Option<StackSetStatus>) -> Self

pub fn template_body(self, input: impl Into<String>) -> Self

The structure that contains the body of the template that was used to create or update the stack set.

pub fn set_template_body(self, input: Option<String>) -> Self

pub fn parameters(self, input: impl Into<Parameter>) -> Self

pub fn set_parameters(self, input: Option<Vec<Parameter>>) -> Self

pub fn capabilities(self, input: impl Into<Capability>) -> Self

pub fn set_capabilities(self, input: Option<Vec<Capability>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn stack_set_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Number (ARN) of the stack set.

pub fn set_stack_set_arn(self, input: Option<String>) -> Self

pub fn administration_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Number (ARN) of the IAM role used to create or update the stack set.

Use customized administrator roles to control which users or groups can manage specific stack sets within the same administrator account. For more information, see Prerequisites: Granting Permissions for Stack Set Operations in the AWS CloudFormation User Guide.

pub fn set_administration_role_arn(self, input: Option<String>) -> Self

pub fn execution_role_name(self, input: impl Into<String>) -> Self

The name of the IAM execution role used to create or update the stack set.

Use customized execution roles to control which stack resources users and groups can include in their stack sets.

pub fn set_execution_role_name(self, input: Option<String>) -> Self

pub fn stack_set_drift_detection_details(
self,
input: StackSetDriftDetectionDetails
) -> Self

Detailed information about the drift status of the stack set.

For stack sets, contains information about the last completed drift operation performed on the stack set. Information about drift operations currently in progress is not included.

pub fn set_stack_set_drift_detection_details(
self,
input: Option<StackSetDriftDetectionDetails>
) -> Self

pub fn auto_deployment(self, input: AutoDeployment) -> Self

[Service-managed permissions] Describes whether StackSets automatically deploys to AWS Organizations accounts that are added to a target organization or organizational unit (OU).

pub fn set_auto_deployment(self, input: Option<AutoDeployment>) -> Self

pub fn permission_model(self, input: PermissionModels) -> Self

Describes how the IAM roles required for stack set operations are created.

With self-managed permissions, you must create the administrator and execution roles required to deploy to target accounts. For more information, see Grant Self-Managed Stack Set Permissions.

With service-managed permissions, StackSets automatically creates the IAM roles required to deploy to accounts managed by AWS Organizations. For more information, see Grant Service-Managed Stack Set Permissions.

pub fn set_permission_model(self, input: Option<PermissionModels>) -> Self

pub fn organizational_unit_ids(self, input: impl Into<String>) -> Self

pub fn set_organizational_unit_ids(self, input: Option<Vec<String>>) -> Self

pub fn build(self) -> StackSet

Consumes the builder and constructs a StackSet