Module aws_sdk_cloudformation::operation

Structs

ActivateType

Activates a public third-party extension, making it available for use in stack templates. For more information, see Using public extensions in the CloudFormation User Guide.

BatchDescribeTypeConfigurations

Returns configuration data for the specified CloudFormation extensions, from the CloudFormation registry for the account and region.

CancelUpdateStack

Cancels an update on the specified stack. If the call completes successfully, the

ContinueUpdateRollback

For a specified stack that is in the UPDATE_ROLLBACK_FAILED state,

CreateChangeSet

Creates a list of changes that will be applied to a stack so that you can review the

CreateStack

Creates a stack as specified in the template. After the call completes successfully,

CreateStackInstances

Creates stack instances for the specified accounts, within the specified Regions. A

CreateStackSet

Creates a stack set.

DeactivateType

Deactivates a public extension that was previously activated in this account and region.

DeleteChangeSet

Deletes the specified change set. Deleting change sets ensures that no one executes

DeleteStack

Deletes a specified stack. Once the call completes successfully, stack deletion

DeleteStackInstances

Deletes stack instances for the specified accounts, in the specified Regions.

DeleteStackSet

Deletes a stack set. Before you can delete a stack set, all of its member stack

DeregisterType

Marks an extension or extension version as DEPRECATED in the CloudFormation registry, removing it from active use. Deprecated extensions or extension versions cannot be used in CloudFormation operations.

DescribeAccountLimits

Retrieves your account's AWS CloudFormation limits, such as the maximum number of

DescribeChangeSet

Returns the inputs for the change set and a list of changes that AWS CloudFormation

DescribePublisher

Returns information about a CloudFormation extension publisher.

DescribeStackDriftDetectionStatus

Returns information about a stack drift detection operation. A stack drift detection

DescribeStackEvents

Returns all stack related events for a specified stack in reverse chronological

DescribeStackInstance

Returns the stack instance that's associated with the specified stack set, AWS

DescribeStackResource

Returns a description of the specified resource in the specified stack.

DescribeStackResourceDrifts

Returns drift information for the resources that have been checked for drift in the

DescribeStackResources

Returns AWS resource descriptions for running and deleted stacks. If

DescribeStackSet

Returns the description of the specified stack set.

DescribeStackSetOperation

Returns the description of the specified stack set operation.

DescribeStacks

Returns the description for the specified stack; if no stack name was specified, then

DescribeType

Returns detailed information about an extension that has been registered.

DescribeTypeRegistration

Returns information about an extension's registration, including its current status and type and version identifiers.

DetectStackDrift

Detects whether a stack's actual configuration differs, or has

DetectStackResourceDrift

Returns information about whether a resource's actual configuration differs, or has

DetectStackSetDrift

Detect drift on a stack set. When CloudFormation performs drift detection on a stack

EstimateTemplateCost

Returns the estimated monthly cost of a template. The return value is an AWS Simple

ExecuteChangeSet

Updates a stack using the input information that was provided when the specified

GetStackPolicy

Returns the stack policy for a specified stack. If a stack doesn't have a policy, a

GetTemplate

Returns the template body for a specified stack. You can get the template for running

GetTemplateSummary

Returns information about a new or existing template. The

ListChangeSets

Returns the ID and status of each active change set for a stack. For example, AWS

ListExports

Lists all exported output values in the account and Region in which you call this

ListImports

Lists all stacks that are importing an exported output value. To modify or remove an

ListStackInstances

Returns summary information about stack instances that are associated with the

ListStackResources

Returns descriptions of all resources of the specified stack.

ListStackSetOperationResults

Returns summary information about the results of a stack set operation.

ListStackSetOperations

Returns summary information about operations performed on a stack set.

ListStackSets

Returns summary information about stack sets that are associated with the

ListStacks

Returns the summary information for stacks whose status matches the specified

ListTypeRegistrations

Returns a list of registration tokens for the specified extension(s).

ListTypeVersions

Returns summary information about the versions of an extension.

ListTypes

Returns summary information about extension that have been registered with CloudFormation.

PublishType

Publishes the specified extension to the CloudFormation registry as a public extension in this region. Public extensions are available for use by all CloudFormation users. For more information on publishing extensions, see Publishing extensions to make them available for public use in the CloudFormation CLI User Guide.

RecordHandlerProgress

Reports progress of a resource handler to CloudFormation.

RegisterPublisher

Registers your account as a publisher of public extensions in the CloudFormation registry. Public extensions are available for use by all CloudFormation users. This publisher ID applies to your account in all AWS regions.

RegisterType

Registers an extension with the CloudFormation service. Registering an extension makes it available for use in CloudFormation templates in your AWS account, and includes:

SetStackPolicy

Sets a stack policy for a specified stack.

SetTypeConfiguration

Specifies the configuration data for a registered CloudFormation extension, in the given account and region.

SetTypeDefaultVersion

Specify the default version of an extension. The default version of an extension will be used in CloudFormation operations.

SignalResource

Sends a signal to the specified resource with a success or failure status. You can

StopStackSetOperation

Stops an in-progress operation on a stack set and its associated stack instances.

TestType

Tests a registered extension to make sure it meets all necessary requirements for being published in the CloudFormation registry.

UpdateStack

Updates a stack as specified in the template. After the call completes successfully,

UpdateStackInstances

Updates the parameter values for stack instances for the specified accounts, within

UpdateStackSet

Updates the stack set, and associated stack instances in the specified accounts and

UpdateTerminationProtection

Updates termination protection for the specified stack. If a user attempts to delete

ValidateTemplate

Validates a specified template. AWS CloudFormation first checks if the template is