Struct aws_sdk_cloudformation::output::describe_change_set_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DescribeChangeSetOutput

Implementations

impl Builder

pub fn change_set_name(self, input: impl Into<String>) -> Self

The name of the change set.

pub fn set_change_set_name(self, input: Option<String>) -> Self

pub fn change_set_id(self, input: impl Into<String>) -> Self

The ARN of the change set.

pub fn set_change_set_id(self, input: Option<String>) -> Self

pub fn stack_id(self, input: impl Into<String>) -> Self

The ARN of the stack that is associated with the change set.

pub fn set_stack_id(self, input: Option<String>) -> Self

pub fn stack_name(self, input: impl Into<String>) -> Self

The name of the stack that is associated with the change set.

pub fn set_stack_name(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

Information about the change set.

pub fn set_description(self, input: Option<String>) -> Self

pub fn parameters(self, input: impl Into<Parameter>) -> Self

pub fn set_parameters(self, input: Option<Vec<Parameter>>) -> Self

pub fn creation_time(self, input: Instant) -> Self

The start time when the change set was created, in UTC.

pub fn set_creation_time(self, input: Option<Instant>) -> Self

pub fn execution_status(self, input: ExecutionStatus) -> Self

If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because AWS CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

pub fn set_execution_status(self, input: Option<ExecutionStatus>) -> Self

pub fn status(self, input: ChangeSetStatus) -> Self

The current status of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

pub fn set_status(self, input: Option<ChangeSetStatus>) -> Self

pub fn status_reason(self, input: impl Into<String>) -> Self

A description of the change set's status. For example, if your attempt to create a change set failed, AWS CloudFormation shows the error message.

pub fn set_status_reason(self, input: Option<String>) -> Self

pub fn notification_ar_ns(self, input: impl Into<String>) -> Self

pub fn set_notification_ar_ns(self, input: Option<Vec<String>>) -> Self

pub fn rollback_configuration(self, input: RollbackConfiguration) -> Self

The rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

pub fn set_rollback_configuration(
self,
input: Option<RollbackConfiguration>
) -> Self

pub fn capabilities(self, input: impl Into<Capability>) -> Self

pub fn set_capabilities(self, input: Option<Vec<Capability>>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn changes(self, input: impl Into<Change>) -> Self

pub fn set_changes(self, input: Option<Vec<Change>>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn include_nested_stacks(self, input: bool) -> Self

Verifies if IncludeNestedStacks is set to True.

pub fn set_include_nested_stacks(self, input: Option<bool>) -> Self

pub fn parent_change_set_id(self, input: impl Into<String>) -> Self

Specifies the change set ID of the parent change set in the current nested change set hierarchy.

pub fn set_parent_change_set_id(self, input: Option<String>) -> Self

pub fn root_change_set_id(self, input: impl Into<String>) -> Self

Specifies the change set ID of the root change set in the current nested change set hierarchy.

pub fn set_root_change_set_id(self, input: Option<String>) -> Self

pub fn build(self) -> DescribeChangeSetOutput

Consumes the builder and constructs a DescribeChangeSetOutput