Struct aws_sdk_cloudformation::output::describe_type_output::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DescribeTypeOutput

Implementations

impl Builder

pub fn arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the extension.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn type(self, input: RegistryType) -> Self

The kind of extension.

pub fn set_type(self, input: Option<RegistryType>) -> Self

pub fn type_name(self, input: impl Into<String>) -> Self

The name of the extension.

If the extension is a public third-party type you have activated with a type name alias, CloudFormation returns the type name alias. For more information, see ActivateType.

pub fn set_type_name(self, input: Option<String>) -> Self

pub fn default_version_id(self, input: impl Into<String>) -> Self

The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

pub fn set_default_version_id(self, input: Option<String>) -> Self

pub fn is_default_version(self, input: bool) -> Self

Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

pub fn set_is_default_version(self, input: Option<bool>) -> Self

pub fn type_tests_status(self, input: TypeTestsStatus) -> Self

The contract test status of the registered extension version. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

PASSED: The extension has passed all its contract tests.

An extension must have a test status of PASSED before it can be published. For more information, see Publishing extensions to make them available for public use in the CloudFormation Command Line Interface User Guide.

FAILED: The extension has failed one or more contract tests.

IN_PROGRESS: Contract tests are currently being performed on the extension.

NOT_TESTED: Contract tests have not been performed on the extension.

pub fn set_type_tests_status(self, input: Option<TypeTestsStatus>) -> Self

pub fn type_tests_status_description(self, input: impl Into<String>) -> Self

The description of the test status. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

pub fn set_type_tests_status_description(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description of the extension.

pub fn set_description(self, input: Option<String>) -> Self

pub fn schema(self, input: impl Into<String>) -> Self

The schema that defines the extension.

For more information on extension schemas, see Resource Provider Schema in the CloudFormation CLI User Guide.

pub fn set_schema(self, input: Option<String>) -> Self

pub fn provisioning_type(self, input: ProvisioningType) -> Self

For resource type extensions, the provisioning behavior of the resource type. AWS CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include all of the following handlers, and therefore cannot actually be provisioned.

create

read

delete

pub fn set_provisioning_type(self, input: Option<ProvisioningType>) -> Self

pub fn deprecated_status(self, input: DeprecatedStatus) -> Self

The deprecation status of the extension version.

Valid values include:

LIVE: The extension is activated or registered and can be used in CloudFormation operations, dependent on its provisioning behavior and visibility scope.

DEPRECATED: The extension has been deactivated or deregistered and can no longer be used in CloudFormation operations.

For public third-party extensions, CloudFormation returns null.

pub fn set_deprecated_status(self, input: Option<DeprecatedStatus>) -> Self

pub fn logging_config(self, input: LoggingConfig) -> Self

Contains logging configuration information for private extensions. This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

pub fn set_logging_config(self, input: Option<LoggingConfig>) -> Self

pub fn required_activated_types(
self,
input: impl Into<RequiredActivatedType>
) -> Self

pub fn set_required_activated_types(
self,
input: Option<Vec<RequiredActivatedType>>
) -> Self

pub fn execution_role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the IAM execution role used to register the extension. This applies only to private extensions you have registered in your account. For more information, see RegisterType.

If the registered extension calls any AWS APIs, you must create an IAM execution role that includes the necessary permissions to call those AWS APIs, and provision that execution role in your account. CloudFormation then assumes that execution role to provide your extension with the appropriate credentials.

pub fn set_execution_role_arn(self, input: Option<String>) -> Self

pub fn visibility(self, input: Visibility) -> Self

The scope at which the extension is visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: The extension is only visible and usable within the account in which it is registered. AWS CloudFormation marks any extensions you register as PRIVATE.

PUBLIC: The extension is publically visible and usable within any Amazon account.

pub fn set_visibility(self, input: Option<Visibility>) -> Self

pub fn source_url(self, input: impl Into<String>) -> Self

The URL of the source code for the extension.

pub fn set_source_url(self, input: Option<String>) -> Self

pub fn documentation_url(self, input: impl Into<String>) -> Self

The URL of a page providing detailed documentation for this extension.

pub fn set_documentation_url(self, input: Option<String>) -> Self

pub fn last_updated(self, input: Instant) -> Self

When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

pub fn set_last_updated(self, input: Option<Instant>) -> Self

pub fn time_created(self, input: Instant) -> Self

When the specified private extension version was registered or activated in your account.

pub fn set_time_created(self, input: Option<Instant>) -> Self

pub fn configuration_schema(self, input: impl Into<String>) -> Self

A JSON string that represent the current configuration data for the extension in this account and region.

To set the configuration data for an extension, use SetTypeConfiguration. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

pub fn set_configuration_schema(self, input: Option<String>) -> Self

pub fn publisher_id(self, input: impl Into<String>) -> Self

The publisher ID of the extension publisher.

This applies only to public third-party extensions. For private registered extensions, and extensions provided by Amazon, CloudFormation returns null.

pub fn set_publisher_id(self, input: Option<String>) -> Self

pub fn original_type_name(self, input: impl Into<String>) -> Self

For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

pub fn set_original_type_name(self, input: Option<String>) -> Self

pub fn original_type_arn(self, input: impl Into<String>) -> Self

For public extensions that have been activated for this account and region, the Amazon Resource Name (ARN) of the public extension.

pub fn set_original_type_arn(self, input: Option<String>) -> Self

pub fn public_version_number(self, input: impl Into<String>) -> Self

The version number of a public third-party extension.

This applies only if you specify a public extension you have activated in your account, or specify a public extension without specifying a version. For all other extensions, CloudFormation returns null.

pub fn set_public_version_number(self, input: Option<String>) -> Self

pub fn latest_public_version(self, input: impl Into<String>) -> Self

The latest version of a public extension that is available for use.

This only applies if you specify a public extension, and you do not specify a version. For all other requests, CloudFormation returns null.

pub fn set_latest_public_version(self, input: Option<String>) -> Self

pub fn is_activated(self, input: bool) -> Self

Whether or not the extension is activated in the account and region.

This only applies to public third-party extensions. For all other extensions, CloudFormation returns null.

pub fn set_is_activated(self, input: Option<bool>) -> Self

pub fn auto_update(self, input: bool) -> Self

Whether CloudFormation automatically updates the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated. For more information, see Activating public extensions for use in your account in the AWS CloudFormation User Guide.

pub fn set_auto_update(self, input: Option<bool>) -> Self

pub fn build(self) -> DescribeTypeOutput

Consumes the builder and constructs a DescribeTypeOutput