Struct aws_sdk_cloudformation::output::DescribeChangeSetOutput

#[non_exhaustive]
pub struct DescribeChangeSetOutput {

    pub change_set_name: Option<String>,
    pub change_set_id: Option<String>,
    pub stack_id: Option<String>,
    pub stack_name: Option<String>,
    pub description: Option<String>,
    pub parameters: Option<Vec<Parameter>>,
    pub creation_time: Option<Instant>,
    pub execution_status: Option<ExecutionStatus>,
    pub status: Option<ChangeSetStatus>,
    pub status_reason: Option<String>,
    pub notification_ar_ns: Option<Vec<String>>,
    pub rollback_configuration: Option<RollbackConfiguration>,
    pub capabilities: Option<Vec<Capability>>,
    pub tags: Option<Vec<Tag>>,
    pub changes: Option<Vec<Change>>,
    pub next_token: Option<String>,
    pub include_nested_stacks: Option<bool>,
    pub parent_change_set_id: Option<String>,
    pub root_change_set_id: Option<String>,

}

The output for the DescribeChangeSet action.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

change_set_name: Option<String>

The name of the change set.

change_set_id: Option<String>

The ARN of the change set.

stack_id: Option<String>

The ARN of the stack that is associated with the change set.

stack_name: Option<String>

The name of the stack that is associated with the change set.

description: Option<String>

Information about the change set.

parameters: Option<Vec<Parameter>>

A list of Parameter structures that describes the input parameters and their values used to create the change set. For more information, see the Parameter data type.

creation_time: Option<Instant>

The start time when the change set was created, in UTC.

execution_status: Option<ExecutionStatus>

If the change set execution status is AVAILABLE, you can execute the change set. If you can’t execute the change set, the status indicates why. For example, a change set might be in an UNAVAILABLE state because AWS CloudFormation is still creating it or in an OBSOLETE state because the stack was already updated.

status: Option<ChangeSetStatus>

The current status of the change set, such as CREATE_IN_PROGRESS, CREATE_COMPLETE, or FAILED.

status_reason: Option<String>

A description of the change set's status. For example, if your attempt to create a change set failed, AWS CloudFormation shows the error message.

notification_ar_ns: Option<Vec<String>>

The ARNs of the Amazon Simple Notification Service (Amazon SNS) topics that will be associated with the stack if you execute the change set.

rollback_configuration: Option<RollbackConfiguration>

The rollback triggers for AWS CloudFormation to monitor during stack creation and updating operations, and for the specified monitoring period afterwards.

capabilities: Option<Vec<Capability>>

If you execute the change set, the list of capabilities that were explicitly acknowledged when the change set was created.

tags: Option<Vec<Tag>>

If you execute the change set, the tags that will be associated with the stack.

changes: Option<Vec<Change>>

A list of Change structures that describes the resources AWS CloudFormation changes if you execute the change set.

next_token: Option<String>

If the output exceeds 1 MB, a string that identifies the next page of changes. If there is no additional page, this value is null.

include_nested_stacks: Option<bool>

Verifies if IncludeNestedStacks is set to True.

parent_change_set_id: Option<String>

Specifies the change set ID of the parent change set in the current nested change set hierarchy.

root_change_set_id: Option<String>

Specifies the change set ID of the root change set in the current nested change set hierarchy.

Implementations

impl DescribeChangeSetOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeChangeSetOutput