Struct aws_sdk_cloudformation::output::DescribeTypeOutput

#[non_exhaustive]
pub struct DescribeTypeOutput {

    pub arn: Option<String>,
    pub type: Option<RegistryType>,
    pub type_name: Option<String>,
    pub default_version_id: Option<String>,
    pub is_default_version: Option<bool>,
    pub type_tests_status: Option<TypeTestsStatus>,
    pub type_tests_status_description: Option<String>,
    pub description: Option<String>,
    pub schema: Option<String>,
    pub provisioning_type: Option<ProvisioningType>,
    pub deprecated_status: Option<DeprecatedStatus>,
    pub logging_config: Option<LoggingConfig>,
    pub required_activated_types: Option<Vec<RequiredActivatedType>>,
    pub execution_role_arn: Option<String>,
    pub visibility: Option<Visibility>,
    pub source_url: Option<String>,
    pub documentation_url: Option<String>,
    pub last_updated: Option<Instant>,
    pub time_created: Option<Instant>,
    pub configuration_schema: Option<String>,
    pub publisher_id: Option<String>,
    pub original_type_name: Option<String>,
    pub original_type_arn: Option<String>,
    pub public_version_number: Option<String>,
    pub latest_public_version: Option<String>,
    pub is_activated: Option<bool>,
    pub auto_update: Option<bool>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

arn: Option<String>

The Amazon Resource Name (ARN) of the extension.

type: Option<RegistryType>

The kind of extension.

type_name: Option<String>

The name of the extension.

If the extension is a public third-party type you have activated with a type name alias, CloudFormation returns the type name alias. For more information, see ActivateType.

default_version_id: Option<String>

The ID of the default version of the extension. The default version is used when the extension version is not specified.

This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

To set the default version of an extension, use SetTypeDefaultVersion .

is_default_version: Option<bool>

Whether the specified extension version is set as the default version.

This applies only to private extensions you have registered in your account, and extensions published by Amazon. For public third-party extensions, whether or not they are activated in your account, CloudFormation returns null.

type_tests_status: Option<TypeTestsStatus>

The contract test status of the registered extension version. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

PASSED: The extension has passed all its contract tests.

An extension must have a test status of PASSED before it can be published. For more information, see Publishing extensions to make them available for public use in the CloudFormation Command Line Interface User Guide.

FAILED: The extension has failed one or more contract tests.

IN_PROGRESS: Contract tests are currently being performed on the extension.

NOT_TESTED: Contract tests have not been performed on the extension.

type_tests_status_description: Option<String>

The description of the test status. To return the extension test status of a specifc extension version, you must specify VersionId.

This applies only to registered private extension versions. CloudFormation does not return this information for public extensions, whether or not they are activated in your account.

description: Option<String>

The description of the extension.

schema: Option<String>

The schema that defines the extension.

For more information on extension schemas, see Resource Provider Schema in the CloudFormation CLI User Guide.

provisioning_type: Option<ProvisioningType>

For resource type extensions, the provisioning behavior of the resource type. AWS CloudFormation determines the provisioning type during registration, based on the types of handlers in the schema handler package submitted.

Valid values include:

FULLY_MUTABLE: The resource type includes an update handler to process updates to the type during stack update operations.

IMMUTABLE: The resource type does not include an update handler, so the type cannot be updated and must instead be replaced during stack update operations.

NON_PROVISIONABLE: The resource type does not include all of the following handlers, and therefore cannot actually be provisioned.

create

read

delete

deprecated_status: Option<DeprecatedStatus>

The deprecation status of the extension version.

Valid values include:

LIVE: The extension is activated or registered and can be used in CloudFormation operations, dependent on its provisioning behavior and visibility scope.

DEPRECATED: The extension has been deactivated or deregistered and can no longer be used in CloudFormation operations.

For public third-party extensions, CloudFormation returns null.

logging_config: Option<LoggingConfig>

Contains logging configuration information for private extensions. This applies only to private extensions you have registered in your account. For public extensions, both those provided by Amazon and published by third parties, CloudFormation returns null. For more information, see RegisterType.

required_activated_types: Option<Vec<RequiredActivatedType>>

For extensions that are modules, the public third-party extensions that must be activated in your account in order for the module itself to be activated.

execution_role_arn: Option<String>

The Amazon Resource Name (ARN) of the IAM execution role used to register the extension. This applies only to private extensions you have registered in your account. For more information, see RegisterType.

If the registered extension calls any AWS APIs, you must create an IAM execution role that includes the necessary permissions to call those AWS APIs, and provision that execution role in your account. CloudFormation then assumes that execution role to provide your extension with the appropriate credentials.

visibility: Option<Visibility>

The scope at which the extension is visible and usable in CloudFormation operations.

Valid values include:

PRIVATE: The extension is only visible and usable within the account in which it is registered. AWS CloudFormation marks any extensions you register as PRIVATE.

PUBLIC: The extension is publically visible and usable within any Amazon account.

source_url: Option<String>

The URL of the source code for the extension.

documentation_url: Option<String>

The URL of a page providing detailed documentation for this extension.

last_updated: Option<Instant>

When the specified extension version was registered. This applies only to:

Private extensions you have registered in your account. For more information, see RegisterType.

Public extensions you have activated in your account with auto-update specified. For more information, see ActivateType.

time_created: Option<Instant>

When the specified private extension version was registered or activated in your account.

configuration_schema: Option<String>

A JSON string that represent the current configuration data for the extension in this account and region.

To set the configuration data for an extension, use SetTypeConfiguration. For more information, see Configuring extensions at the account level in the CloudFormation User Guide.

publisher_id: Option<String>

The publisher ID of the extension publisher.

This applies only to public third-party extensions. For private registered extensions, and extensions provided by Amazon, CloudFormation returns null.

original_type_name: Option<String>

For public extensions that have been activated for this account and region, the type name of the public extension.

If you specified a TypeNameAlias when enabling the extension in this account and region, CloudFormation treats that alias as the extension's type name within the account and region, not the type name of the public extension. For more information, see Specifying aliases to refer to extensions in the CloudFormation User Guide.

original_type_arn: Option<String>

For public extensions that have been activated for this account and region, the Amazon Resource Name (ARN) of the public extension.

public_version_number: Option<String>

The version number of a public third-party extension.

This applies only if you specify a public extension you have activated in your account, or specify a public extension without specifying a version. For all other extensions, CloudFormation returns null.

latest_public_version: Option<String>

The latest version of a public extension that is available for use.

This only applies if you specify a public extension, and you do not specify a version. For all other requests, CloudFormation returns null.

is_activated: Option<bool>

Whether or not the extension is activated in the account and region.

This only applies to public third-party extensions. For all other extensions, CloudFormation returns null.

auto_update: Option<bool>

Whether CloudFormation automatically updates the extension in this account and region when a new minor version is published by the extension publisher. Major versions released by the publisher must be manually updated. For more information, see Activating public extensions for use in your account in the AWS CloudFormation User Guide.

Implementations

impl DescribeTypeOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeTypeOutput