Module aws_sdk_cloudformation::error

Modules

already_exists_error

See AlreadyExistsError

cfn_registry_error

See CFNRegistryError

change_set_not_found_error

See ChangeSetNotFoundError

created_but_modified_error

See CreatedButModifiedError

insufficient_capabilities_error

See InsufficientCapabilitiesError

invalid_change_set_status_error

See InvalidChangeSetStatusError

invalid_operation_error

See InvalidOperationError

invalid_state_transition_error

See InvalidStateTransitionError

limit_exceeded_error

See LimitExceededError

name_already_exists_error

See NameAlreadyExistsError

operation_id_already_exists_error

See OperationIdAlreadyExistsError

operation_in_progress_error

See OperationInProgressError

operation_not_found_error

See OperationNotFoundError

operation_status_check_failed_error

See OperationStatusCheckFailedError

stack_instance_not_found_error

See StackInstanceNotFoundError

stack_set_not_empty_error

See StackSetNotEmptyError

stack_set_not_found_error

See StackSetNotFoundError

stale_request_error

See StaleRequestError

token_already_exists_error

See TokenAlreadyExistsError

type_not_found_error

See TypeNotFoundError

Structs

AlreadyExistsError

The resource with the name requested already exists.

CFNRegistryError

An error occurred during a CloudFormation registry operation.

CancelUpdateStackError

ChangeSetNotFoundError

The specified change set name or ID doesn't exit. To view valid change sets for a

ContinueUpdateRollbackError

CreateChangeSetError

CreateStackError

CreateStackInstancesError

CreateStackSetError

CreatedButModifiedError

The specified resource exists, but has been changed.

DeleteChangeSetError

DeleteStackError

DeleteStackInstancesError

DeleteStackSetError

DeregisterTypeError

DescribeAccountLimitsError

DescribeChangeSetError

DescribeStackDriftDetectionStatusError

DescribeStackEventsError

DescribeStackInstanceError

DescribeStackResourceDriftsError

DescribeStackResourceError

DescribeStackResourcesError

DescribeStackSetError

DescribeStackSetOperationError

DescribeStacksError

DescribeTypeError

DescribeTypeRegistrationError

DetectStackDriftError

DetectStackResourceDriftError

DetectStackSetDriftError

EstimateTemplateCostError

ExecuteChangeSetError

GetStackPolicyError

GetTemplateError

GetTemplateSummaryError

InsufficientCapabilitiesError

The template contains resources with capabilities that weren't specified in the

InvalidChangeSetStatusError

The specified change set can't be used to update the stack. For example, the change

InvalidOperationError

The specified operation isn't valid.

InvalidStateTransitionError

Error reserved for use by the CloudFormation CLI. CloudFormation does not return this error to users.

LimitExceededError

The quota for the resource has already been reached.

ListChangeSetsError

ListExportsError

ListImportsError

ListStackInstancesError

ListStackResourcesError

ListStackSetOperationResultsError

ListStackSetOperationsError

ListStackSetsError

ListStacksError

ListTypeRegistrationsError

ListTypeVersionsError

ListTypesError

NameAlreadyExistsError

The specified name is already in use.

OperationIdAlreadyExistsError

The specified operation ID already exists.

OperationInProgressError

Another operation is currently in progress for this stack set. Only one operation can

OperationNotFoundError

The specified ID refers to an operation that doesn't exist.

OperationStatusCheckFailedError

Error reserved for use by the CloudFormation CLI. CloudFormation does not return this error to users.

RecordHandlerProgressError

RegisterTypeError

SetStackPolicyError

SetTypeDefaultVersionError

SignalResourceError

StackInstanceNotFoundError

The specified stack instance doesn't exist.

StackSetNotEmptyError

You can't yet delete this stack set, because it still contains one or more stack

StackSetNotFoundError

The specified stack set doesn't exist.

StaleRequestError

Another operation has been performed on this stack set since the specified operation

StopStackSetOperationError

TokenAlreadyExistsError

A client request token already exists.

TypeNotFoundError

The specified type does not exist in the CloudFormation registry.

UpdateStackError

UpdateStackInstancesError

UpdateStackSetError

UpdateTerminationProtectionError

ValidateTemplateError

Enums

CancelUpdateStackErrorKind

ContinueUpdateRollbackErrorKind

CreateChangeSetErrorKind

CreateStackErrorKind

CreateStackInstancesErrorKind

CreateStackSetErrorKind

DeleteChangeSetErrorKind

DeleteStackErrorKind

DeleteStackInstancesErrorKind

DeleteStackSetErrorKind

DeregisterTypeErrorKind

DescribeAccountLimitsErrorKind

DescribeChangeSetErrorKind

DescribeStackDriftDetectionStatusErrorKind

DescribeStackEventsErrorKind

DescribeStackInstanceErrorKind

DescribeStackResourceDriftsErrorKind

DescribeStackResourceErrorKind

DescribeStackResourcesErrorKind

DescribeStackSetErrorKind

DescribeStackSetOperationErrorKind

DescribeStacksErrorKind

DescribeTypeErrorKind

DescribeTypeRegistrationErrorKind

DetectStackDriftErrorKind

DetectStackResourceDriftErrorKind

DetectStackSetDriftErrorKind

EstimateTemplateCostErrorKind

ExecuteChangeSetErrorKind

GetStackPolicyErrorKind

GetTemplateErrorKind

GetTemplateSummaryErrorKind

ListChangeSetsErrorKind

ListExportsErrorKind

ListImportsErrorKind

ListStackInstancesErrorKind

ListStackResourcesErrorKind

ListStackSetOperationResultsErrorKind

ListStackSetOperationsErrorKind

ListStackSetsErrorKind

ListStacksErrorKind

ListTypeRegistrationsErrorKind

ListTypeVersionsErrorKind

ListTypesErrorKind

RecordHandlerProgressErrorKind

RegisterTypeErrorKind

SetStackPolicyErrorKind

SetTypeDefaultVersionErrorKind

SignalResourceErrorKind

StopStackSetOperationErrorKind

UpdateStackErrorKind

UpdateStackInstancesErrorKind

UpdateStackSetErrorKind

UpdateTerminationProtectionErrorKind

ValidateTemplateErrorKind